requirements.txt
mongoengine
python-dotenv
flask
PyJWT
marshmallow
marshmallow-oneofschema

# dev deps
black
flake8
pep8-naming
pre-commit
